python/requirements.txt
flask
flask-cors
PyPDF2
openai
boto3
python-dotenv
requests
